RW-6035 move unsaved changes check to concept-search component
